Utils: add helper functions to get standard directories
