+++ src/Utils/DirHelpers.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <string>
+
+namespace gamescope
+{
+	std::string_view GetHomeDir();
+	std::string GetLocalUsrDir();
+	std::string GetUsrDir();
+	std::string_view GetConfigDir();
+}
